Refactor the common::executor * Use `wasm_bindgen_futures::spawn_local` instead of custom runtime implementation * Use `js_sys::Date::now` instead of defined in js `date_now` * Use 'setTimeout' within Timer
