+++ Cargo.toml
@@ -27,7 +27,7 @@ path = "mm2src/docker_tests.rs"
 [lib]
 name = "mm2"
 path = "mm2src/mm2_lib.rs"
-crate-type = ["staticlib"]
+crate-type = ["cdylib", "staticlib"]
 test = false
 doctest = false
 bench = false
