+++ wasm_build/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>MM2 example</title>
+</head>
+<body>
+
+</body>
+<script src="script.js"></script>
+</html>
